+++ Lessons_3_4/Task3_Drivers/Factory/DriverFactory.cs
@@ -1,35 +1,36 @@
 ﻿using Bogus;
 using Task3_Drivers.Model;
-using Person = Task3_Drivers.Model.Person;
 
 namespace Task3_Drivers.Factory;
 
-public class DriverFactory
+public class DriverFactory2
 {
     private readonly Faker _faker;
 
-    public DriverFactory()
+    public DriverFactory2()
     {
         _faker = new Faker();
     }
 
-    public Driver Create(Person person)
+    public Driver Create()
     {
-        var driver = new Driver();
+        var driver = new Driver
+        {
+            Name = _faker.Name.FirstName(),
+            Surname = _faker.Name.LastName(),
+            DateOfBirth = GetDateOfBirth(),
+            LicenseID = new Guid()
+        };
 
-        driver.Name = _faker.Name.FirstName();
-        driver.Surname = _faker.Name.LastName();
-        driver.DateOfBirth = person.DateOfBirth;
-        driver.LicenseID = new Guid();
-        driver.DateDriverLicense = ReleaseLicenseDate(person);
+        driver.DateDriverLicense = GetDateDriverLicense(driver.DateOfBirth);
         driver.IsDriver = true;
 
         return driver;
     }
 
-    private DateTime ReleaseLicenseDate( Person person)
+    private DateTime GetDateDriverLicense(DateTime dateOfBirth)
     {
-        DateTime startDate = person.DateOfBirth.AddYears(16);
+        DateTime startDate = dateOfBirth.AddYears(21);
         DateTime finishDate = DateTime.Now;
         DateTime date = _faker.Date.BetweenOffset(
             new DateTimeOffset(startDate), new DateTimeOffset(finishDate)).Date;
@@ -37,20 +38,12 @@ public class DriverFactory
         return date;
     }
 
-    private Driver EstablishLicenseData(DateTime dateOfBirth, Driver license)
+    private DateTime GetDateOfBirth()
     {
-        var faker = new Faker();
+        DateTime date = _faker.Date.BetweenOffset(
+            new DateTimeOffset(DateTime.Now.AddYears(-100)), 
+            new DateTimeOffset(DateTime.Now.AddYears(-23))).Date;
 
-        if (DateTime.Now.Year - dateOfBirth.Year > 16)
-        {
-            license.DateDriverLicense = faker.Date.BetweenOffset(
-                new DateTimeOffset(dateOfBirth), new DateTimeOffset(DateTime.Today)).Date;
-            return license;
-        }
-        else
-        {
-            Console.WriteLine("Person too young");
-            return null;
-        }
+        return date;
     }
 }
