--- Lessons_3_4/Task3_Drivers/Resources/VehicleType.cs
@@ -1,6 +1,0 @@
-﻿namespace Task3_Drivers.Resources;
-
-public enum VehicleType
-{
-    SportCar, Track, Minivan
-}
